old test cases deleted
